Switch entrypoint and main functionality
