+++ README.md
@@ -79,22 +79,19 @@ contents of this directory with your own code.
 There are a few things to keep in mind when writing your action code:
 
 - Most GitHub Actions toolkit and CI/CD operations are processed asynchronously.
-  In `index.ts`, you will see that the action is run in an `async` function.
+  In `main.ts`, you will see that the action is run in an `async` function.
 
   ```javascript
-  import * as core from '@actions/core';
-  ...
+  import * as core from '@actions/core'
+  //...
 
   async function run() {
     try {
-        ...
-    }
-    catch (error) {
-      core.setFailed(error.message);
+      //...
+    } catch (error) {
+      core.setFailed(error.message)
     }
   }
-
-  run()
   ```
 
   For more information about the GitHub Actions toolkit, see the
